remoe simpla-lite, bundle promises for now (to be stripped w/ v1 specs)
