+++ .gitignore
@@ -1,8 +1,6 @@
 # Dist
 simpla.js
 simpla.min.js
-simpla-lite.js
-simpla-lite.min.js
 
 # Deps
 node_modules
